채팅방 나가기 기능 추가

+++ server/src/main/java/com/swjeon/pethealthcaremanager/server/Repository/ChatRoomRepository.java
@@ -3,5 +3,5 @@ package com.swjeon.pethealthcaremanager.server.Repository;
 import com.swjeon.pethealthcaremanager.server.Entity.ChatRoomEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ChatRoomRepository extends JpaRepository<ChatRoomEntity, Long> {
+public interface ChatRoomRepository extends JpaRepository<ChatRoomEntity, Integer> {
 }
